divide the main script into smaller scripts
